footer + nav fix
