+++ frontend/src/components/Footer/Footer.css
@@ -6,10 +6,9 @@
 	margin-left: -7px;
 	margin-right: -7px; */
 	justify-content: space-evenly;
-	max-height: 300px;
 	bottom: 0;
-	padding: 1em;
-    background: none;
+    padding: 0.8em;
+    background: linear-gradient(to right, lightblue, plum);
 }
 
 .fab {
@@ -30,11 +29,11 @@
 .githubLink,
 .porfolio-link {
 	text-decoration: none;
-	font-size: 2em;
+	font-size: 1.2em;
 }
 
 .about-me-div {
-	font-size: 3em;
+	font-size: 2em;
 	color: white;
 	text-shadow: -1px 3px 3px black, 3px 3px 3px black, 3px -1px 0 black,
 		-1px -1px 0 black;
