+++ frontend/src/components/MainPage/MainPage.css
@@ -1,17 +1,18 @@
 .main-page-div{
      background-image: url('../../project-items/wallpapersden.com_valkyries-magic-the-gathering_3840x2160.jpg');
     /* position: absolute; */
-    top: 0;
-    left: 0;
+    /* top: 0;
+    left: 0; */
     overflow: auto;
     width: 100vw;
-    height: 100vh;
+    height: 95vh;
     background-repeat: no-repeat;
     background-position: center;
     background-size: cover;
     display: inline-block;   
     /* opacity: 0.6; 
     filter: alpha(opacity=60); */
+    
 }
 
 .card-name-div{
